[FIX] Fix data type definition in Windows designs for 64 bit systems
The data type definitions in the driver.h, drvintf.h and drvintf.c
of Windows designs are revised in order to comply with different
bit sizes of the target system (e.g. 64 bit).

+++ stack/include/common/driver.h
@@ -92,8 +92,8 @@ kernel stack and mapped into user virtual address space.
 */
 typedef struct
 {
-    UINT32                  memSize;        ///< Size of PDO to be allocated and mapped
-    UINT32                  pdoMemOffset;   ///< Offset of PDO memory returned by kernel
+    size_t                  memSize;        ///< Size of PDO to be allocated and mapped
+    size_t                  pdoMemOffset;   ///< Offset of PDO memory returned by kernel
 } tPdoMem;
 
 /**
@@ -118,7 +118,7 @@ typedef struct
 {
     void*                   pKernelAddr;    ///< Pointer to the Kernel address
     void*                   pUserAddr;      ///< Pointer to the User address
-    UINT32                  size;           ///< Size of the shared memory
+    size_t                  size;           ///< Size of the shared memory
 } tMemStruc;
 
 //------------------------------------------------------------------------------
